+++ sim/firesim-lib/src/main/cc/bridges/tracerv.cc
@@ -255,7 +255,7 @@ void tracerv_t::serialize(
     const bool fireperf) {
   const int max_consider = std::min(max_core_ipc, 7);
   if (human_readable || test_output) {
-    for (int i = 0; i < (bytes_received / sizeof(uint64_t)); i += 8) {
+    for (uint64_t i = 0; i < (bytes_received / sizeof(uint64_t)); i += 8) {
       if (test_output) {
         fprintf(tracefile, "%016lx", OUTBUF[i + 7]);
         fprintf(tracefile, "%016lx", OUTBUF[i + 6]);
@@ -282,7 +282,7 @@ void tracerv_t::serialize(
     }
   } else if (fireperf) {
 
-    for (int i = 0; i < (bytes_received / sizeof(uint64_t)); i += 8) {
+    for (uint64_t i = 0; i < (bytes_received / sizeof(uint64_t)); i += 8) {
       uint64_t cycle_internal = OUTBUF[i + 0];
 
       for (int q = 0; q < max_consider; q++) {
@@ -298,11 +298,11 @@ void tracerv_t::serialize(
       }
     }
   } else {
-    for (int i = 0; i < (bytes_received / sizeof(uint64_t)); i += 8) {
+    for (uint64_t i = 0; i < (bytes_received / sizeof(uint64_t)); i += 8) {
       // this stores as raw binary. stored as little endian.
       // e.g. to get the same thing as the human readable above,
       // flip all the bytes in each 512-bit line.
-      for (int q = 0; q < 8; q++) {
+      for (uint64_t q = 0; q < 8; q++) {
         fwrite(OUTBUF + (i + q), sizeof(uint64_t), 1, tracefile);
       }
     }
